feat(contracts): 옵션 C TrendClusterSummary post_count int → float (γ 자연 후속)
scope: γ (PR #18) 의 input 측 (ScoreHistory + ClusterScoringContext) post_count int→float
마이그 후 출력 contract (TrendClusterSummary + ClusterDecision) 가 int 잔존이라 drift.
share-weighted fan-out (β2/β3/β4) 의 fractional mass 가 자연스럽게 출력에 보존되도록 정합.

핵심 변경:
- contracts/output.py: TrendClusterSummary.post_count_total/today int → float.
- aggregation/build_cluster_summary.py: ClusterDecision 동일 필드 float.
- pipelines/run_scoring_pipeline.py: round() 제거 (γ round 제거 정책 확장).
  classify_lifecycle 의 int cast 제거.
- scoring/direction.py: classify_lifecycle(post_count_total: int → float) signature.
  early threshold 비교는 부동소수 비교로 자연 동작.
- tests/snapshot/golden_summaries.json: 11 entries × 2 필드 = 22 entries 갱신
  (1 → 1.0, JSON dump 표기 변경).
- sample_data/contract_fixtures/trend_cluster_summary.json: 1 fixture 갱신.

무관 (확인됨):
- StarRocks DDL — post_count_* 컬럼 부재 (representative_weekly 는 effective_item_count
  + total_item_contribution 기반). DDL 마이그 불필요.
- backend_poster / render_canonical_smoke_html.py — post_count 참조 없음.

테스트: 773 passed regression 0 (Pydantic int → float coerce 자동 호환).

상세: memory/project_option_c_completion_2026_04_29.md

+++ src/aggregation/build_cluster_summary.py
@@ -52,8 +52,9 @@ class ClusterDecision:
     lifecycle_stage: LifecycleStage
     data_maturity: DataMaturity
     display_name: str
-    post_count_total: int
-    post_count_today: int
+    # 옵션 C (2026-04-29): share-weighted fan-out (β2/β3/β4) 으로 fractional.
+    post_count_total: float
+    post_count_today: float
     avg_engagement_rate: float
     total_video_views: int
     top_video_ids: list[str]
